Add "partial interest" concept for `interesttarget`
This adds the "partial interest" concept for `interesttarget`, activated
via keyboard focus. Focusing an interest invoker shows interest, but
in "partial interest" mode. Hovering or focusing the target element
upgrades to "full interest" mode. As of this CL, there is no difference
between the modes, but subsequent CLs will limit keyboard focus in
partial interest mode.

In the process, this CL also cleans up and unifies the logic for
hover/dehover with the logic for focus/blur, so that they work almost
identically.

Bug: 326681249
Change-Id: Ia246688d03f0f38d341d15a2d1a0811ffcded13c